+++ package.json
@@ -91,12 +91,6 @@
       ]
     }
   },
-  "dependencies": {
-    "@graphql-markdown/docusaurus": "file:packages/docusaurus",
-    "@graphql-markdown/core": "file:packages/core",
-    "@graphql-markdown/diff": "file:packages/diff",
-    "@graphql-markdown/printer-legacy": "file:packages/printer-legacy"
-  },
   "devDependencies": {
     "@graphql-markdown/tools-config": "file:config",
     "@stryker-mutator/core": "6.2.3",
