+++ packages/diff/package.json
@@ -61,7 +61,8 @@
   },
   "dependencies": {
     "@graphql-inspector/core": "^3.4.0",
-    "@graphql-tools/load": "^7.7.7"
+    "@graphql-tools/load": "^7.7.7",
+    "@graphql-markdown/utils": "1.14.0"
   },
   "peerDependencies": {
     "@graphql-tools/graphql-file-loader": "^7.5.5"
