+++ packages/diff/src/index.js
@@ -6,7 +6,9 @@ const { loadSchema } = require("@graphql-tools/load");
 const { diff } = require("@graphql-inspector/core");
 const { GraphQLFileLoader } = require("@graphql-tools/graphql-file-loader");
 
-const { fileExists, readFile, saveFile } = require("./utils/helpers/fs");
+const {
+  fs: { fileExists, readFile, saveFile },
+} = require("@graphql-markdown/utils").helpers;
 
 const SCHEMA_HASH_FILE = ".schema";
 const SCHEMA_REF = "schema.graphql";
