+++ packages/utils/.npmignore
@@ -1,0 +1,14 @@
+scripts
+tests
+output
+reports
+coverage
+build
+.nyc_output
+
+.dccache
+
+jest.config.js
+jest.setup.js
+nyc.config.js
+stryket.conf.js
