+++ packages/utils/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require("@graphql-markdown/tools-config").jest;
