+++ packages/utils/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@graphql-markdown/core",
+  "name": "@graphql-markdown/utils",
   "homepage": "https://graphql-markdown.github.io",
-  "description": "GraphQL-Markdown core package for generating Markdown documentation from a GraphQL schema.",
+  "description": "GraphQL-Markdown common libraries.",
   "bugs": {
     "url": "https://github.com/graphql-markdown/graphql-markdown/issues"
   },
@@ -13,6 +13,7 @@
     "url": "https://github.com/graphql-markdown/graphql-markdown.git"
   },
   "keywords": [
+    "docusaurus",
     "graphql",
     "graphql-schema",
     "documentation",
@@ -60,26 +61,14 @@
     "stryker": "stryker run"
   },
   "dependencies": {
-    "@graphql-markdown/utils": "1.14.0",
-    "@graphql-markdown/printer-legacy": "1.14.0"
+    "@graphql-tools/load": "^7.7.7"
   },
   "peerDependencies": {
-    "@graphql-markdown/diff": "^1.14.0",
-    "prettier": "^2.5"
-  },
-  "peerDependenciesMeta": {
-    "@graphql-markdown/diff": {
-      "optional": true
-    },
-    "prettier": {
-      "optional": true
-    }
+    "graphql": "^14.0 || ^15.0 || ^16.0"
   },
   "devDependencies": {
     "@graphql-tools/graphql-file-loader": "7.5.5",
-    "@graphql-markdown/diff": "1.14.0",
     "@graphql-markdown/tools-config": "1.0.0",
-    "@stryker-mutator/core": "6.2.3",
     "@stryker-mutator/jest-runner": "6.2.3",
     "diff": "5.1.0",
     "eslint": "8.25.0",
