+++ packages/utils/src/scalars/string.js
