+++ packages/utils/stryker.conf.js
@@ -1,0 +1,1 @@
+module.exports = require("@graphql-markdown/tools-config").stryker;
